2023.8.3 okr老版本

+++ src/redux/reducer/common.ts
@@ -1,10 +1,11 @@
-import { actionTypes as commonActionTypes } from '../actions/commonActions';
+import { getSearchParamValue } from "../../services/util";
+import { actionTypes as commonActionTypes } from "../actions/commonActions";
 interface Message {
   visible: boolean;
   text: string;
-  messageType: 'success' | 'info' | 'warning' | 'error' | undefined;
+  messageType: "success" | "info" | "warning" | "error" | undefined;
 }
-
+declare var window: Window;
 export interface Common {
   loading: boolean;
   showApps: boolean;
@@ -15,6 +16,7 @@ export interface Common {
   moveState: string;
   chatState: boolean;
   showChatState: boolean;
+  createMemberState: boolean;
   unChatNum: number | string;
   unMessageNum: number;
   socketObj: any;
@@ -24,9 +26,16 @@ export interface Common {
   taskMemberVisible: boolean;
   taskMemberX: number;
   taskMemberY: number;
+  taskCustomVisible: boolean;
+  taskCustomX: number;
+  taskCustomY: number;
   fileInfo: any;
   fileKey: string;
   fileVisible: boolean;
+  inviteKey: string;
+  animateState: boolean;
+  contentVisible: boolean;
+  colorState: boolean;
 }
 
 const defaultState: Common = {
@@ -36,14 +45,15 @@ const defaultState: Common = {
   showSearch: false,
   message: {
     visible: false,
-    text: '',
+    text: "",
     messageType: undefined,
   },
   // headerIndex: 3,
   headerIndex: null,
-  moveState: '',
+  moveState: "",
   chatState: false,
   showChatState: false,
+  createMemberState: false,
   unChatNum: 0,
   unMessageNum: 0,
   socketObj: null,
@@ -53,19 +63,31 @@ const defaultState: Common = {
   taskMemberVisible: false,
   taskMemberX: 0,
   taskMemberY: 0,
+  taskCustomVisible: false,
+  taskCustomX: 0,
+  taskCustomY: 0,
   fileInfo: null,
-  fileKey: '',
+  fileKey: "",
   fileVisible: false,
+  inviteKey: "",
+  animateState: false,
+  contentVisible: true,
+  colorState: false,
 };
-
+const deviceType = getSearchParamValue(window.location.search, "deviceType");
 export const common = (state = defaultState, action: any) => {
   switch (action.type) {
     case commonActionTypes.FAILED:
-      if (action.error.statusCode === '701') {
-        const redirect = `${window.location.protocol}//${window.location.host}`;
-        // window.location.href = `https://account.qingtime.cn?apphigh=27&redirect=${redirect}&logo=https://working.vip/page/logo2.svg`;
-        localStorage.clear();
-        window.location.href = `${redirect}/`;
+      if (action.error.statusCode === "701") {
+        if (deviceType !== "tool") {
+          const redirect = `${window.location.protocol}//${window.location.host}`;
+          // window.location.href = `https://account.qingtime.cn?apphigh=27&redirect=${redirect}&logo=https://working.vip/page/logo2.svg`;
+          localStorage.clear();
+          window.open(`${redirect}/welcome`, "_self");
+          state.inviteKey = "";
+        } else {
+          alert("请在chrome插件里重新登录");
+        }
       }
       return {
         ...state,
@@ -73,8 +95,9 @@ export const common = (state = defaultState, action: any) => {
         message: {
           visible: true,
           text: action.error.msg,
-          severity: 'error',
+          severity: "error",
         },
+        inviteKey: state.inviteKey,
       };
     case commonActionTypes.SWITCH_APPS:
       return {
@@ -132,6 +155,11 @@ export const common = (state = defaultState, action: any) => {
         ...state,
         showChatState: action.showChatState,
       };
+    case commonActionTypes.SET_CREATEMEMBER_CHATSTATE:
+      return {
+        ...state,
+        createMemberState: action.createMemberState,
+      };
     case commonActionTypes.SET_UNCHATREADNUM:
       return {
         ...state,
@@ -161,10 +189,21 @@ export const common = (state = defaultState, action: any) => {
         taskMemberX: action.taskMemberX,
         taskMemberY: action.taskMemberY,
       };
+    case commonActionTypes.CHANGE_TASKCUSTOM_VISIBLE:
+      return {
+        ...state,
+        taskCustomVisible: action.taskCustomVisible,
+        taskCustomX: action.taskCustomX,
+        taskCustomY: action.taskCustomY,
+      };
     case commonActionTypes.SET_FILEINFO:
       return {
         ...state,
         fileInfo: action.fileInfo,
+      };
+    case commonActionTypes.SET_FILEVISIBLE:
+      return {
+        ...state,
         fileVisible: action.fileVisible,
       };
     case commonActionTypes.SET_FILEKEY:
@@ -172,7 +211,26 @@ export const common = (state = defaultState, action: any) => {
         ...state,
         fileKey: action.fileKey,
       };
-
+    case commonActionTypes.SET_INVITEKEY:
+      return {
+        ...state,
+        inviteKey: action.inviteKey,
+      };
+    case commonActionTypes.CHANGE_ANIMATESTATE:
+      return {
+        ...state,
+        animateState: action.animateState,
+      };
+    case commonActionTypes.CHANGE_CONTENTVISIBLE:
+      return {
+        ...state,
+        contentVisible: action.contentVisible,
+      };
+    case commonActionTypes.CHANGE_COLORSTATE:
+      return {
+        ...state,
+        colorState: action.colorState,
+      };
     default:
       return state;
   }
